MAking changes to  destination properties

+++ src/test/java/tz/go/moh/him/thscp/mediator/irims/main/MediatorMainTest.java
@@ -4,7 +4,6 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.openhim.mediator.engine.MediatorConfig;
 import tz.go.moh.him.thscp.mediator.irims.MediatorMain;
-
 import java.lang.reflect.Method;
 
 public class MediatorMainTest {
@@ -23,7 +22,7 @@ public class MediatorMainTest {
         MediatorConfig mediatorConfig = (MediatorConfig) loadConfigMethod.invoke(null, "src/test/resources/mediator.properties");
 
         Assert.assertEquals("localhost", mediatorConfig.getServerHost());
-        Assert.assertEquals(new Integer(3020), mediatorConfig.getServerPort());
+        Assert.assertEquals(new Integer(3024), mediatorConfig.getServerPort());
         Assert.assertEquals(new Integer(60000), mediatorConfig.getRootTimeout());
         Assert.assertTrue(mediatorConfig.getHeartsbeatEnabled());
     }
